Updated coaches layout
Updated SCSS files and the style sheet. The style sheet should compile
correctly. Did some clean up. The custom field method calls are not
working. I didn't do anything with that, just noticed they weren't
working.

There also is no longer a need to add a p with a class of coachDesc to
the coaches post. Please remove that from the posts.